+++ src/screens/transaction_details/components/index.ts
@@ -1,9 +1,9 @@
 import Overview from './overview';
-import Messages from './messages';
 import Logs from './logs';
+import Instructions from './instructions';
 
 export {
   Overview,
-  Messages,
   Logs,
+  Instructions,
 };