+++ src/screens/transaction_details/types.ts
@@ -8,9 +8,19 @@ export type OverviewType = {
 
 export type LogType = string;
 
+export type InstructionType = {
+  type: string;
+  index: number;
+  innerIndex: number;
+  program: string;
+  value: any;
+  rawData: string;
+}
+
 export type TransactionState = {
   loading: boolean;
   exists: boolean;
   overview: OverviewType;
   logs: LogType[];
+  instructions: InstructionType[];
 }