+++ src/screens/validator_details/components/transactions/index.test.tsx
@@ -12,7 +12,6 @@ import Transactions from '.';
 jest.mock('@components', () => ({
   Box: (props) => <div id="Box" {...props} />,
   TransactionsList: (props) => <div id="TransactionsList" {...props} />,
-  TransactionListDetails: (props) => <div id="TransactionListDetails" {...props} />,
 }));
 
 // ==================================