+++ CHANGELOG.md
@@ -22,4 +22,6 @@
 - Update general chain config report issue url ([\#98](https://github.com/forbole/big-dipper-2.0-solana/issues/98))
 - Fix AvatarName if name === '""' or length is empty return address ([\#95](https://github.com/forbole/big-dipper-2.0-solana/issues/95))
 - Add tx details logs component
+- Add tx details instructions ([\#105](https://github.com/forbole/big-dipper-2.0-solana/issues/105))
 - Fix tabs index behavior (validators page) ([\#69](https://github.com/forbole/big-dipper-2.0-solana/issues/69))
+